Moved main Wikipedia table extraction code to tablextract library, improved accuracy.
