+++ environment.yml
@@ -155,6 +155,7 @@ dependencies:
     - sphinx-autobuild==0.7.1
     - sphinxcontrib-napoleon==0.7
     - sphinxcontrib-websupport==1.1.0
+    - tablextract>=1.1.7
     - tblib==1.3.2
     - termcolor==1.1.0
     - terminado==0.8.1
@@ -177,7 +178,7 @@ dependencies:
     - werkzeug==0.14.1
     - wget==3.2
     - widgetsnbextension==3.4.2
-    - wikipedia-api==0.3.7
+    - wikipedia-api==0.4.4
     - wrapt==1.10.11
     - xlrd==1.1.0
     - xlwt==1.3.0
