+++ requirements.txt
@@ -60,6 +60,7 @@ Send2Trash==1.5.0
 simplegeneric==0.8.1
 six==1.11.0
 spacy==2.0.17
+tablextract>=1.1.7
 termcolor==1.1.0
 terminado==0.8.1
 testpath==0.4.2
@@ -70,7 +71,7 @@ urllib3==1.23
 wcwidth==0.1.7
 webencodings==0.5.1
 Werkzeug==0.14.1
-Wikipedia-API==0.3.7
+wikipedia-api==0.4.4
 wget==3.2
 widgetsnbextension==3.4.2
 xlrd==1.1.0
